+++ site-cookbooks/stack-java/attributes/default.rb
@@ -1,0 +1,5 @@
+default['java']['jdk_version'] = '8'
+default['java']['remove_deprecated_packages'] = true
+default['java']['oracle']['accept_oracle_download_terms'] = true
+default['java']['java_home'] = '/usr/lib/java/default'
+default['java']['set_default'] = false