+++ site-cookbooks/stack-java/metadata.rb
@@ -1,10 +1,9 @@
-name             'stack-marathon'
+name             'stack-java'
 maintainer       'The Startup Stack'
 maintainer_email 'avi@avi.io'
 license          'MIT'
-description      'Installs/Configures stack-marathon'
+description      'Installs/Configures stack-java'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.19'
+version          '0.1.1'
 
-depends 'apt'
-depends 'stack-java'
+depends 'java'