+++ site-cookbooks/stack-base/CHANGELOG.md
@@ -1,7 +1,10 @@
 # The Startup-Stack (base) changelog
 
+## 1.0.4
+
+* [@KensoDev] - 2016/02/15 - Remove 'vim' cookbook dependency
+* Adding 'vim' to be installed as a package
+
 ## 1.0.0
 
 * [@KensoDev] - 2015/08/04 - First release production ready base
-
-